Hybrid Core beta1

+++ library/admin/meta-box-post-seo.php
@@ -15,6 +15,7 @@
 
 /* Add the post SEO meta box on the 'add_meta_boxes' hook. */
 add_action( 'add_meta_boxes', 'hybrid_meta_box_post_add_seo', 10, 2 );
+add_action( 'add_meta_boxes', 'hybrid_meta_box_post_remove_seo', 10, 2 );
 
 /* Save the post SEO meta box data on the 'save_post' hook. */
 add_action( 'save_post', 'hybrid_meta_box_post_save_seo', 10, 2 );
@@ -23,6 +24,8 @@ add_action( 'save_post', 'hybrid_meta_box_post_save_seo', 10, 2 );
  * Adds the post SEO meta box for all public post types.
  *
  * @since 1.2.0
+ * @param string $post_type The post type of the current post being edited.
+ * @param object $post The current post being edited.
  * @return void
  */
 function hybrid_meta_box_post_add_seo( $post_type, $post ) {
@@ -32,6 +35,25 @@ function hybrid_meta_box_post_add_seo( $post_type, $post ) {
 	/* Only add meta box if current user can edit, add, or delete meta for the post. */
 	if ( ( true === $post_type_object->public ) && ( current_user_can( 'edit_post_meta', $post->ID ) || current_user_can( 'add_post_meta', $post->ID ) || current_user_can( 'delete_post_meta', $post->ID ) ) )
 		add_meta_box( 'hybrid-core-post-seo', __( 'SEO', 'hybrid-core' ), 'hybrid_meta_box_post_display_seo', $post_type, 'normal', 'high' );
+}
+
+/**
+ * Remove the meta box from some post types.
+ *
+ * @since 1.3.0
+ * @param string $post_type The post type of the current post being edited.
+ * @param object $post The current post being edited.
+ * @return void
+ */ 
+function hybrid_meta_box_post_remove_seo( $post_type, $post ) {
+
+	/* Removes post stylesheets support of the bbPress 'topic' post type. */
+	if ( function_exists( 'bbp_get_topic_post_type' ) && bbp_get_topic_post_type() == $post_type )
+		remove_meta_box( 'hybrid-core-post-seo', bbp_get_topic_post_type(), 'normal' );
+
+	/* Removes post stylesheets support of the bbPress 'reply' post type. */
+	elseif ( function_exists( 'bbp_get_reply_post_type' ) && bbp_get_reply_post_type() == $post_type )
+		remove_meta_box( 'hybrid-core-post-seo', bbp_get_reply_post_type(), 'normal' );
 }
 
 /**
@@ -78,14 +100,10 @@ function hybrid_meta_box_post_save_seo( $post_id, $post ) {
 	if ( !isset( $_POST['hybrid-core-post-seo'] ) || !wp_verify_nonce( $_POST['hybrid-core-post-seo'], basename( __FILE__ ) ) )
 		return $post_id;
 
-	/* Check if the current user has permission to edit the post meta. */
-	if ( !current_user_can( 'edit_post_meta', $post_id ) )
-		return $post_id;
-
 	$meta = array(
-		'Title' => 	strip_tags( $_POST['hybrid-document-title'] ),
-		'Description' => 	strip_tags( $_POST['hybrid-meta-description'] ),
-		'Keywords' => 	strip_tags( $_POST['hybrid-meta-keywords'] )
+		'Title' => 	$_POST['hybrid-document-title'],
+		'Description' => 	$_POST['hybrid-meta-description'],
+		'Keywords' => 	$_POST['hybrid-meta-keywords']
 	);
 
 	foreach ( $meta as $meta_key => $new_meta_value ) {
@@ -94,11 +112,15 @@ function hybrid_meta_box_post_save_seo( $post_id, $post ) {
 		$meta_value = get_post_meta( $post_id, $meta_key, true );
 
 		/* If there is no new meta value but an old value exists, delete it. */
-		if ( '' == $new_meta_value && $meta_value )
+		if ( current_user_can( 'delete_post_meta', $post_id, $meta_key ) && '' == $new_meta_value && $meta_value )
 			delete_post_meta( $post_id, $meta_key, $meta_value );
 
+		/* If a new meta value was added and there was no previous value, add it. */
+		elseif ( current_user_can( 'add_post_meta', $post_id, $meta_key ) && $new_meta_value && '' == $meta_value )
+			add_post_meta( $post_id, $meta_key, $new_meta_value, true );
+
 		/* If the new meta value does not match the old value, update it. */
-		elseif ( $new_meta_value && $new_meta_value != $meta_value )
+		elseif ( current_user_can( 'edit_post_meta', $post_id, $meta_key ) && $new_meta_value && $new_meta_value != $meta_value )
 			update_post_meta( $post_id, $meta_key, $new_meta_value );
 	}
 }
